tweak C Makefile and example
