+++ pgml-sdks/pgml/c/Makefile
@@ -1,31 +1,17 @@
 BINARY_NAME=pgml
+HEADER=include/${BINARY_NAME}.h
+PGML_LIB=../target/debug/
 
-build:
-	cargo build --features c
-	cargo expand --features c > expanded.rs
-	cbindgen --lang C -o pgml.h expanded.rs
-	# GOARCH=amd64 GOOS=darwin go build -o ${BINARY_NAME}-darwin main.go
-	GOARCH=amd64 GOOS=linux go build -o ${BINARY_NAME}-linux pgml.go
-	# GOARCH=amd64 GOOS=windows go build -o ${BINARY_NAME}-windows main.go
+bindings:
+	cargo b --features c
+	RUST_TOOLCHAIN=$(dirname $(rustup +nightly which rustc)) cbindgen --config cbindgen.toml --output ${HEADER} ../
 
-build_test:
-	cargo build --features c
-	cargo expand --features c > expanded.rs
-	cbindgen --lang C -o pgml.h expanded.rs
-	gcc test.c -o test -l pgml -L ./../target/debug
-
-test: build_test
-	LD_LIBRARY_PATH=./../target/debug ./test
-
-test_c:
-	gcc test.c -o test -l pgml -L ./../target/debug
-	LD_LIBRARY_PATH=./../target/debug ./test
+build: bindings
+	gcc -Wall -o build/example -Iinclude/ -L${PGML_LIB} -l ${BINARY_NAME} example/main.c
 
 run: build
-	LD_LIBRARY_PATH=./../target/debug ./${BINARY_NAME}-linux
+	LD_LIBRARY_PATH=${PGML_LIB} ./build/example
 
 clean:
-	go clean
-	# rm ${BINARY_NAME}-darwin
-	rm ${BINARY_NAME}-linux
-	# rm ${BINARY_NAME}-windows
+	rm ${HEADER}
+
