--- pgml-sdks/pgml/c/go/go.mod
@@ -1,3 +1,0 @@
-module pgml
-
-go 1.22.2
